dashboards/saedn/requirements.txt
dash    
plotly
psycopg2-binary
numpy
pandas
python-dotenv
